Remove template filename
The generated fail2ban.conf used to contain the full path to the
template used for generating the file on the Puppet Master.  This was
made uneffective by the switch to epp templates.

Because it used to display the full path of the template on the Pullet
Master, it produces more clutter than value when switching environment
for testing changes (each environment changes the content of
fail2ban.conf).

For this reason, drop this feature.